rename PlanQK-Service-IA to PlanQK-Platform-IA
