+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/files/io_github_ust_quantil_artifacttemplates_PlanQKPlatform__ContainerManagementInterface.war
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fb5f327feb6a6ac32fb1e85d56b0d0ed653915a6b712253bc1c7ea05a0917f77
+size 48242960
