+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/source/README.md
@@ -1,0 +1,62 @@
+# Generated Implementation Artifact Skeleton for io_github_ust_quantil_artifacttemplates_PlanQKPlatform__ContainerManagementInterface
+
+* Interface Name: io_github_ust_quantil_artifacttemplates_PlanQKPlatform__ContainerManagementInterface
+* Namespace: https://ust-quantil.github.io/artifacttemplates
+* Java Package: io.github.ust.quantil.artifacttemplates
+
+## Preconditions
+To develop and build this Implementation Artifact you need Maven and Eclipse.
+- How to install maven is described at the end of this file
+- If you use an Eclipse workspace for the first time with a maven project you've to run this command before starting eclipse:
+  `mvn -Declipse.workspace="<your_workspace_location>" eclipse:configure-workspace`
+  - When using Windows: Make sure you replace `\` (backslashes) with `/` (slashes)
+  - It does not hurt to run it multiple times
+  - This command must not be executed in a particular folder
+
+## Preparations
+- Unzip the archive generated by Winery into a location of your choice.
+- Note: All following maven commands are executed on a shell in the root folder of the Implementation Artifact.
+
+## Create Eclipse project
+- Run `mvn eclipse:eclipse`
+- Open Eclipse
+- 'File' -> 'Import' -> 'Existing Project into Workspace' -> click 'Next'
+- Click 'Browse' and select folder where you unzipped the generated Implementation Artifact
+- Select the project from the list and click 'Finish'
+
+## Test your Implementation Artifact
+- Run `mvn clean package tomcat7:run-war`
+ (more information: http://tomcat.apache.org/maven-plugin-trunk/tomcat7-maven-plugin/plugin-info.html)
+- Open this page to see the list of available services: http://localhost:9090/services/
+- This page also links the WSDL
+- With the WSDL you are able to test your IA using SOAPui or other tools.
+
+### Upload your Implementation Artifact
+You have to options to do this:
+
+### 1) Automatically (to the Winery instance this IA project was generated with)
+- Run `mvn deploy`
+- The WAR is directly uploaded into the correct ArtifactTemplate. Previous versions are overwritten.
+
+### 2) Manually
+- Run `mvn clean package`
+- Locate the WAR file in the /target folder
+- Open Winery in your browser, locate the NodeType and then click through to the Implementation of the respective Interface, its Implementation Artifacts and lastly to the ArtifactTemplate representing this IA.
+- Upload the WAR file by clicking on the "Attach file..." button
+
+## Install Maven
+When using [chocolatey](https://chocolatey.org/): Just run `choco install maven`.
+We recommend to install JDK8 via chocolatey, too: `choco install jdk8`.
+This ensures all required settings are in place (e.g., `JAVA_HOME` always uptodate).
+
+In case you want to install it automatically:
+- Download latest release of Maven: http://maven.apache.org/download.cgi (Binary zip)
+- Unzip into directory of your choice
+- Set environment variables (in system environment variables, not the ones for the user account):
+  - Check that the variable `JAVA_HOME` is set. If not, create it and enter the path to your JDK, e.g., `C:\Program Files\Java\jdk1.8.0_112`.
+  - Create the following environment variables
+    - `M2_HOME`: `C:\Path\where\your\maven\was\unzipped\to` (no backslash at the end)
+    - `M2`: `%M2_HOME%\bin`
+    - `MAVEN_OPTS`: `-Xms256m -Xmx512m`
+	- Append `;%M2%` to the `Path` environment variable. Make sure to separate the new entry with a semicolon (`;`) from the last one.
+- After that you have to open a new command line windows (if you already opened one) to let the changes take effect.
