+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/source/ia/src/main/java/io/github/ust/quantil/artifacttemplates/WsdlFilter.java
