+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/source/ia/src/main/java/io/github/ust/quantil/artifacttemplates/planqkplatform/PlanQKPlatformConstants.java
@@ -1,0 +1,13 @@
+package io.github.ust.quantil.artifacttemplates.planqkplatform;
+
+public abstract class PlanQKPlatformConstants {
+
+    // Namespace under which the SOAP service operates
+    public static final String NAMESPACE_URI = "https://ust-quantil.github.io/artifacttemplates";
+
+    // Port type to use for the SOAP service
+    public static final String PORT_TYPE_NAME = "io_github_ust_quantil_artifacttemplates_PlanQKPlatform__ContainerManagementInterfacePort";
+
+    // Name of the XML schema file to use as basis for the WSDL generation
+    public static final String XSD_NAME = "containerManagementInterface.xsd";
+}
