--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/ArtifactTemplate.tosca
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<Definitions xmlns="http://docs.oasis-open.org/tosca/ns/2011/12" xmlns:winery="http://www.opentosca.org/winery/extensions/tosca/2013/02/12" xmlns:selfservice="http://www.eclipse.org/winery/model/selfservice" xmlns:testwineryopentoscaorg="http://test.winery.opentosca.org" targetNamespace="https://ust-quantil.github.io/artifacttemplates" id="artifactTemplatesQuantil-PlanQK-Service-IA_w1-wip1">
-    <ArtifactTemplate xmlns:otatyIgeneral="http://opentosca.org/artifacttypes" name="PlanQK-Service-IA_w1-wip1" type="otatyIgeneral:WAR-Java17" id="PlanQK-Service-IA_w1-wip1">
-        <Properties>
-            <WSProperties xmlns="http://www.uni-stuttgart.de/opentosca">
-                <ServiceEndpoint>/io_github_ust_quantil_artifacttemplates_PlanQKService__ContainerManagementInterfacePort</ServiceEndpoint>
-                <PortType>{http://artifacttemplates.opentosca.org}io_github_ust_quantil_artifacttemplates_PlanQKService__ContainerManagementInterface</PortType>
-                <InvocationType>SOAP/HTTP</InvocationType>
-            </WSProperties>
-        </Properties>
-        <ArtifactReferences>
-            <ArtifactReference reference="artifacttemplates/https%253A%252F%252Fust-quantil.github.io%252Fartifacttemplates/PlanQK-Service-IA_w1-wip1/files/io_github_ust_quantil_artifacttemplates_PlanQKService__ContainerManagementInterface.war"/>
-        </ArtifactReferences>
-    </ArtifactTemplate>
-</Definitions>
