--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/files/io_github_ust_quantil_artifacttemplates_PlanQKService__ContainerManagementInterface.war
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:639f153d1b8d0e35665f774c65eed6c330907b7af11b4e2642c8081c709c5abf
-size 48242927
