--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/source/ia/src/test/java/io/github/ust/quantil/artifacttemplates/planqkservice/PlanQKServiceManagementInterfaceEndpointIntegrationTest.java
@@ -1,32 +1,0 @@
-package io.github.ust.quantil.artifacttemplates.planqkservice;
-
-import java.io.IOException;
-
-import io.github.ust.quantil.artifacttemplates.OpenToscaIASpringApplication;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-@SpringBootTest(classes = OpenToscaIASpringApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class PlanQKServiceManagementInterfaceEndpointIntegrationTest {
-
-    @LocalServerPort
-    private int serverPort;
-
-    @Test
-    void getWsdlUsingQueryParameter() throws IOException {
-        HttpGet request = new HttpGet(getBasePath() + "?wsdl");
-        CloseableHttpResponse response = HttpClientBuilder.create().build().execute(request);
-
-        assertEquals(200, response.getStatusLine().getStatusCode());
-    }
-
-    private String getBasePath() {
-        return "http://localhost:" + serverPort + "/" + PlanQKServiceConstants.PORT_TYPE_NAME;
-    }
-}
